% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM_IHS_Parallel.R
\name{IHS_BackTransform}
\alias{IHS_BackTransform}
\title{Back transform dataset with IHS}
\usage{
IHS_BackTransform(dat, lambda, lambdaType = 1, cluster = 1)
}
\arguments{
\item{dat}{Dataset to be back transformed. Must be numeric.}

\item{lambda}{Transformation parameters.}

\item{lambdaType}{Dimensions of lambda matrix. 1 means p-dimensional lambda vector. 2 means (kxp)-dimensional lambda matrix. 1 is default.}
}
\value{
Matrix containing observations on the transformed scale.
\itemize{
\item objectiveScore Value of objective function.
\item transWSS WSS on transformed scale
\item cluster Cluster assignmets
\item centers Cluster centers. On pre-transformation scale.
\item lambda Estimated lambda values for IHS transformation.
\item iter Number of iterations made by TiK-means
\item size Cluster sizes
\item lambdaType lambdaType specified in function call
\item stepType lambdaStepType specified in function call
}
}
\description{
Converts a dataset to its transformed counterpart using parameters returned from tikmeans.
}
