% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM_IHS_Parallel.R
\name{tikmeans}
\alias{tikmeans}
\title{TiK-means: Transformation-infused K-means}
\usage{
tikmeans(dat, K, centers = NULL, lambdaType = 1, lambdaSeq = seq(0,
  10, length.out = 51), lambdaStepType = 1, lambda = NULL,
  nstart = 100, maxiter = 1000, numCores = detectCores()\%/\%2,
  verbose = FALSE)
}
\arguments{
\item{dat}{Dataset to be clustered. Must be numeric.}

\item{K}{Number of clusters to estimate.}

\item{centers}{Initial cluster means. Must have K rows.}

\item{lambdaType}{Dimensions of lambda matrix. 1 means p-dimensional lambda vector. 2 means (kxp)-dimensional lambda matrix. 1 is default.}

\item{lambdaSeq}{Lambda search grid. Passed to `tikmeans` function}

\item{lambdaStepType}{Type of step made on lambda. 1 is default and suggested.}

\item{lambda}{Initial values of the lambda matrix.}

\item{nstart}{Number of initial points for tikmeans.}

\item{maxiter}{Max number of iterations in tikmeans.}

\item{numCores}{Number of cores used in parallel computation.}

\item{verbose}{Whether to print details of initial points}
}
\value{
List containing results from TiK-means.
\itemize{
\item objectiveScore Value of objective function.
\item transWSS WSS on transformed scale
\item cluster Cluster assignmets
\item centers Cluster centers. On pre-transformation scale.
\item lambda Estimated lambda values for IHS transformation.
\item iter Number of iterations made by TiK-means
\item size Cluster sizes
\item lambdaType lambdaType specified in function call
\item stepType lambdaStepType specified in function call
}
}
\description{
Performs K-means clustering and estimates Inverse Hyperbolic Sine transformation for the provided data and K value. Default values are provided, but setting them based on data is common.
}
