% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM_IHS_Parallel.R
\name{tikmeans_KSelect}
\alias{tikmeans_KSelect}
\title{TiK-means: Transformation-infused K-means K selection algorithm}
\usage{
tikmeans_KSelect(dat, K_min = 1, K_max, exps = seq(0, 100, length.out =
  1000), lambdaSeq = seq(0, 10, length.out = 51), lambdaStart = NULL,
  nstart = 100, maxiter = 10000, lambdaType = 1,
  lambdaStepType = 1, verbose = FALSE,
  numCores = detectCores()\%/\%2)
}
\arguments{
\item{dat}{Dataset to cluster}

\item{K_min}{Lower bound for clusters to check. Default 1.}

\item{K_max}{Upper bound for clusters to check.}

\item{exps}{Exponents used in jump statistic calculations.}

\item{lambdaSeq}{Lambda search grid. Passed to `tikmeans` function}

\item{nstart}{Number of initial points for tikmeans}

\item{maxiter}{Max number of iterations in tikmeans}

\item{lambdaType}{Type of lambda values. Must be 1 or 2. 1 is a single lambda per dimension. 2 is a lambda per dimension*cluster}

\item{lambdaStepType}{Type of step made on lambda}

\item{verbose}{Whether to print details of initial points}

\item{numCores}{Number of cores used in parallel computation.}
}
\value{
List of 3 items
\itemize{
\item K_tikMeans K selected by jump statistic
\item resMat Results of distorted objective function. Each row is calculated with a different exponent from exps.
\item km_list TiK-means results for K = K_min to K = K_max
}
}
\description{
Iterates from K_min to K_max and calculates the jump statistic over various values of $\\eta$ (exps). Prints plots, returns matrix of jump statistics and TiK-means results.
}
